Allow to configure HTTP cache directory

+++ src/mkdocs_include_markdown_plugin/cache.py
@@ -9,9 +9,6 @@ import time
 from importlib.util import find_spec
 
 
-CACHE_AVAILABLE = find_spec('platformdirs') is not None
-
-
 class Cache:
     """Cache for arbitrary content, one file per entry."""
 
@@ -64,31 +61,52 @@ class Cache:
     def clean(self) -> None:
         """Clean expired entries from the cache."""
         for fname in os.listdir(self.cache_dir):
+            if fname == '.gitignore':
+                continue
             fpath = os.path.join(self.cache_dir, fname)
             creation_time = self.get_creation_time_from_fpath(fpath)
             if time.time() > creation_time + self.expiration_seconds:
                 os.remove(fpath)
 
 
-def get_cache_directory() -> str | None:
-    """Get the cache directory."""
-    if not CACHE_AVAILABLE:
+def get_cache_directory(cache_dir: str) -> str | None:
+    """Get cache directory."""
+    if cache_dir:
+        return cache_dir
+
+    if not is_platformdirs_installed():
         return None
 
     try:
         from platformdirs import user_data_dir
     except ImportError:
         return None
-
-    cache_dir = user_data_dir('mkdocs-include-markdown-plugin')
-    os.makedirs(cache_dir, exist_ok=True)
-
-    return cache_dir
+    else:
+        return user_data_dir('mkdocs-include-markdown-plugin')
 
 
-def initialize_cache(expiration_seconds: int) -> Cache | None:
+def initialize_cache(expiration_seconds: int, cache_dir: str) -> Cache | None:
     """Initialize a cache instance."""
-    cache_dir = get_cache_directory()
-    return None if cache_dir is None else Cache(
-        cache_dir, expiration_seconds,
-    )
+    cache_directory = get_cache_directory(cache_dir)
+
+    if cache_directory is None:
+        return None
+
+    os.makedirs(cache_directory, exist_ok=True)
+
+    # Add a `.gitignore` file to prevent the cache directory from being
+    # included in the repository. This is needed because the cache directory
+    # can be configured as a relative path with `cache_dir` setting.
+    gitignore = os.path.join(cache_directory, '.gitignore')
+    if not os.path.exists(gitignore):
+        with open(gitignore, 'wb') as f:
+            f.write(b'*\n')
+
+    cache = Cache(cache_directory, expiration_seconds)
+    cache.clean()
+    return cache
+
+
+def is_platformdirs_installed() -> bool:
+    """Check if `platformdirs` package is installed without importing it."""
+    return find_spec('platformdirs') is not None
